requirements.txt
discord
flask
thefuzz
requests
httpx
pubchempy
pillow
privatebinapi